% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess Text for Natural Language Processing}
\usage{
preprocess(
  text,
  expand_contractions = TRUE,
  extract_words = TRUE,
  remove_stopwords = TRUE,
  lemmatize_words = TRUE
)
}
\arguments{
\item{text}{Input text to be preprocessed.}

\item{expand_contractions}{Logical, indicating whether to expand contractions.}

\item{extract_words}{Logical, indicating whether to tokenize and extract words.}

\item{remove_stopwords}{Logical, indicating whether to remove stopwords.}

\item{lemmatize_words}{Logical, indicating whether to lemmatize words.}
}
\value{
Preprocessed text based on the specified steps.
}
\description{
This function applies a series of text preprocessing steps to enhance the quality
of input text for natural language processing tasks. The steps include expanding
contractions, extracting meaningful words, removing common English stopwords,
and lemmatizing words.
}
\examples{
# Example usage:
preprocessed_text <- preprocess("It's a sample text with contractions.",
                                expand_contractions = TRUE,
                                extract_words = TRUE,
                                remove_stopwords = TRUE,
                                lemmatize_words = TRUE)
print(preprocessed_text)

}
